smoke test + files reorganization
